Refactor task creation and export
